Add files via upload
Website allows for creating a user and a login page for existing users, a store for adding items to a cart receiving a total for all purchases, users can find out information about the course as well as update their user status as well. The backend allows for admins with privileges to edit certain user accounts and player accounts ring up items and delete and insert users into the system as well. Users and players can be viewed. Inventory of items sold can also be viewed by users with admin privileges.